+++ sdk/storage/azure-storage-queue/azure/storage/queue/_queue_client.py
@@ -629,7 +629,7 @@ class QueueClient(StorageAccountHostsMixin):
                 cls=self._config.message_decode_policy,
                 **kwargs
             )
-            return ItemPaged(command, results_per_page=messages_per_page, page_iterator_class=MessagesPaged, messages_to_retrieve=max_messages)
+            return ItemPaged(command, results_per_page=messages_per_page, page_iterator_class=MessagesPaged, max_messages=max_messages)
         except HttpResponseError as error:
             process_storage_error(error)